First draft of the NPCs for blackjack and the tavern
Still a lot to do. Currently the quest can not be started, but at least you can have a tavern brawl. The dialogues have to be created and the NPCs stats adjusted, etc.
The blackjack game will be unplayable until the ATCS code for editng the random requirments gets merged.